requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==1.26.18
python-dotenv==1.0.0
textstat==0.7.3 